requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.30.0
PyPDF2==3.0.1
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
httpx==0.25.2
cachetools==5.3.2
numpy==1.24.3 